feat(travis): try splitting the tests - no issue
