+++ .travis.yml
@@ -32,13 +32,22 @@ cache:
 matrix:
   include:
     # Nodejs - Webdriver.io
-    - env: TEST_SUITE=test:functional
+    - env:
+       - TEST_SUITE=test:functional
+       - TEST_BROWSER=firefox
       addons:
         apt:
           packages:
             - graphicsmagick
         sauce_connect: true
-
+    - env:
+       - TEST_SUITE=test:functional
+       - TEST_BROWSER=chrome
+      addons:
+        apt:
+          packages:
+            - graphicsmagick
+        sauce_connect: true
     # Nodejs - Lint
     - env: TEST_SUITE=test:coding-conventions
