+++ test/utils/travis.js
@@ -1,0 +1,1 @@
+module.exports.isTravis = 'TRAVIS' in process.env && 'CI' in process.env;
